requirements.txt
httpx>=0.27.0
python-whois>=0.8.0
rich>=13.7.1
